fix(server): cast COUNT(*) to integer for PostgreSQL compatibility
PostgreSQL COUNT(*) returns bigint which the pg driver delivers as
a string. The Schema.Number decode fails on strings. Cast to integer.

+++ apps/server/src/orchestration/Layers/ProjectionSnapshotQuery.ts
@@ -656,8 +656,8 @@ const makeProjectionSnapshotQuery = Effect.gen(function* () {
     execute: () =>
       sql`
         SELECT
-          (SELECT COUNT(*) FROM projection_projects) AS "projectCount",
-          (SELECT COUNT(*) FROM projection_threads) AS "threadCount"
+          (SELECT COUNT(*)::integer FROM projection_projects) AS "projectCount",
+          (SELECT COUNT(*)::integer FROM projection_threads) AS "threadCount"
       `,
   });
